% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost}
\alias{cost}
\title{Calculate the test cost given the true test labels and the predicted ones.}
\usage{
cost(y, ypred, beta, A, form = c("average", "total"))
}
\arguments{
\item{y}{the true labels of the test data}

\item{ypred}{the predicted labels of the test data}

\item{beta}{the cost ratio of FP and FN cases. Default = 0.5.}

\item{A}{the cost of FN cases.}

\item{form}{the form of the test cost. If \code{"average"}, the average cost will be calculated. If \code{total}, the total cost on the test set will be calculated.}
}
\value{
the predicted labels.
}
\description{
Calculate the test cost given the true test labels and the predicted ones.
}
\references{
Tian, Y., & Zhang, W. (2018). THORS: An Efficient Approach for Making Classifiers Cost-sensitive. arXiv preprint arXiv:1811.028
}
\seealso{
\code{\link{thors}}, \code{\link{cost}}.
}
