% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.thors.R
\name{predict.thors}
\alias{predict.thors}
\title{Predict the outcome of a set of new observations using the fitted thors object.}
\usage{
\method{predict}{thors}(object, newx = NULL, ...)
}
\arguments{
\item{object}{the fitted npmc object using \code{thors}.}

\item{newx}{a set of new observations.}

\item{...}{additional arguments.}
}
\value{
the test cost.
}
\description{
Predict the outcome of a set of new observations using the fitted thors object.
}
\seealso{
\code{\link{thors}}, \code{\link{cost}}.
}
